Fixed unintentional maximum `vram_texture_preload_size`
Previously, `vram_texture_preload_size` was unintentionally limited
below 4096 MB. This was caused by using `size_t` for internal
calculations, which is 32-bit here.

+++ common/config.hpp
@@ -30,7 +30,7 @@ namespace jkgm {
         std::string command = "jk.exe";
         std::string data_path = "jkgm";
         std::optional<std::string> log_path;
-        size_t vram_texture_preload_size = 0U;
+        uint64_t vram_texture_preload_size = 0U;
     };
 
     std::unique_ptr<config> load_config_file();
